refactor submit restaurant form
